[TASK] Remove hard dependency to typo3/cms-form
Relates: #644
