+++ Classes/Hooks/ItemsProcFunc.php
@@ -8,23 +8,18 @@ namespace Extcode\Cart\Hooks;
  * For the full copyright and license information, please read the
  * LICENSE file that was distributed with this source code.
  */
+
 use Extcode\Cart\Utility\TemplateLayout;
 use TYPO3\CMS\Backend\Utility\BackendUtility as BackendUtilityCore;
 use TYPO3\CMS\Core\Localization\LanguageService;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
-use TYPO3\CMS\Extbase\Configuration\ConfigurationManagerInterface;
-use TYPO3\CMS\Form\Mvc\Configuration\ConfigurationManagerInterface as ExtFormConfigurationManagerInterface;
-use TYPO3\CMS\Form\Mvc\Persistence\FormPersistenceManagerInterface;
 
 class ItemsProcFunc
 {
     protected TemplateLayout $templateLayoutsUtility;
 
-    public function __construct(
-        private readonly FormPersistenceManagerInterface $formPersistenceManager,
-        private readonly ConfigurationManagerInterface $configurationManager,
-        private readonly ExtFormConfigurationManagerInterface $extFormConfigurationManager,
-    ) {
+    public function __construct()
+    {
         $this->templateLayoutsUtility = GeneralUtility::makeInstance(TemplateLayout::class);
     }
 
@@ -33,7 +28,6 @@ class ItemsProcFunc
      */
     public function user_templateLayout(array &$config): void
     {
-        $pageId = 0;
         $currentColPos = $config['flexParentDatabaseRow']['colPos'];
         $pageId = $this->getPageId($config['flexParentDatabaseRow']['pid']);
 
@@ -52,44 +46,6 @@ class ItemsProcFunc
                 array_push($config['items'], $additionalLayout);
             }
         }
-    }
-
-    public function user_formDefinition(array &$config): void
-    {
-        if ($config['field'] !== 'form_definition') {
-            return;
-        }
-
-        $prototypeName = 'cart';
-        if (is_array($config['config']['itemsProcFuncConfig']) && !empty($config['config']['itemsProcFuncConfig']['prototypeName'])) {
-            $prototypeName = $config['config']['itemsProcFuncConfig']['prototypeName'];
-        }
-
-        $formSettings = $this->getFormSettings();
-        $availableForms = $this->formPersistenceManager->listForms($formSettings);
-
-        foreach ($availableForms as $availableForm) {
-            $form = $this->formPersistenceManager->load($availableForm['persistenceIdentifier'], $formSettings, []);
-
-            if ($form['prototypeName'] === $prototypeName) {
-                $config['items'][] = [
-                    0 => $availableForm['name'],
-                    1 => $availableForm['persistenceIdentifier'],
-                ];
-            }
-        }
-    }
-
-    private function getFormSettings(): array
-    {
-        $typoScriptSettings = $this->configurationManager->getConfiguration(ConfigurationManagerInterface::CONFIGURATION_TYPE_SETTINGS, 'form');
-        $formSettings = $this->extFormConfigurationManager->getYamlConfiguration($typoScriptSettings, false);
-        if (!isset($formSettings['formManager'])) {
-            // Config sub array formManager is crucial and should always exist. If it does
-            // not, this indicates an issue in config loading logic. Except in this case.
-            throw new \LogicException('Configuration could not be loaded', 1723717461);
-        }
-        return $formSettings;
     }
 
     /**
@@ -126,8 +82,6 @@ class ItemsProcFunc
      */
     protected function getPageId(int $pid): int
     {
-        $pid = (int)$pid;
-
         if ($pid > 0) {
             return $pid;
         }
