+++ Documentation/Changelog/11.0/Index.rst
@@ -1,6 +1,6 @@
 .. include:: ../../Includes.rst.txt
 
-10.0 Changes
+11.0 Changes
 ============
 
 **Table of contents**
