+++ Documentation/guides.xml
@@ -11,8 +11,8 @@
                interlink-shortcode="extcode/cart"
     />
     <project title="Cart"
-             release="11.2.0"
-             version="11.2"
+             release="11.3.0"
+             version="11.3"
              copyright="2018 - 2025"
     />
     <inventory id="t3tsref" url="https://docs.typo3.org/typo3cms/TyposcriptReference/"/>
