+++ Tests/Functional/EventListener/Mail/AttachmentFromOrderItemTest.php
@@ -24,8 +24,6 @@ class AttachmentFromOrderItemTest extends FunctionalTestCase
 
     public function setUp(): void
     {
-        $this->coreExtensionsToLoad[] = 'typo3/cms-form';
-
         $this->testExtensionsToLoad[] = 'extcode/cart';
         $this->testExtensionsToLoad[] = 'typo3conf/ext/cart/Tests/Fixtures/cart_example';
