+++ ext_emconf.php
@@ -4,7 +4,7 @@ $EM_CONF['cart'] = [
     'title' => 'Cart',
     'description' => 'Shopping Cart(s) for TYPO3',
     'category' => 'plugin',
-    'version' => '11.2.0',
+    'version' => '11.3.0',
     'state' => 'stable',
     'author' => 'Daniel Gohlke',
     'author_email' => 'ext@extco.de',
